FireProjectile() and Attack()
Have them nearly complete. I need to figure out how to get the decals to show up. The template uses them as prefabs I need to learn more. Getting a lot of exceptions when firing in the editor. My first step is getting the decals working and go from there